requirements.txt
asgiref==3.10.0
Django==5.2.8
django-phonenumber-field==8.3.0
phonenumbers==9.0.18
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
